extract base classes for internal and public proxies
